Add files via upload
Here is your task
Your task is to scaffold the frontend of the Augur system. You must create a single-page application that displays a series of metrics for each distributor in the system. Eventually, your frontend will integrate with the backend your team is building, but for now, simply populate it with mock data. Your dashboard should display the following information for each distributor:

Name.
Quantity of goods shipped last month.
Forecasted quantity of goods to ship this coming month.
Year-to-date average goods shipped per month.